Network/Socket/requirements.txt
python-dotenv
requests
httpx
httpx[http2]
